feat: 댓글 기능의 컨트롤러 구현 (#14)
컨트롤러
- `/comment/controller/CommentController`
  - 뉴스와 연관된 댓글 목록 조회
  - 뉴스에 댓글 생성
  - 뉴스에 있는 댓글 삭제

서비스 수정사항
- `/comment/service.CommentService`
- `/comment/service.CommentServiceImpl`
  - 뉴스 삭제 메서드의 매개변수에 newsId 추가

+++ backend/src/main/java/com/tamnara/backend/comment/service/CommentServiceImpl.java
@@ -25,7 +25,7 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public List<CommentDTO> getComments(Long newsId, Integer page, Integer size) {
         if (!newsRepository.existsById(newsId)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "요청하신 뉴스가 존재하지 않습니다.");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "요청하신 댓글의 뉴스가 존재하지 않습니다.");
         }
         Page<Comment> comments = commentRepository.findAllByNewsId(newsId, PageRequest.of(page, size));
 
@@ -45,7 +45,7 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public Long save(Long newsId, CommentCreateRequest commentCreateRequest) {
         if (!newsRepository.existsById(newsId)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "요청하신 뉴스가 존재하지 않습니다.");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "요청하신 댓글의 뉴스가 존재하지 않습니다.");
         }
 
         Comment comment = new Comment();
@@ -54,7 +54,11 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public Long delete(Long commentId) {
+    public Long delete(Long newsId, Long commentId) {
+        if (!newsRepository.existsById(newsId)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "요청하신 댓글의 뉴스가 존재하지 않습니다.");
+        }
+
         commentRepository.deleteById(commentId);
         return commentId;
     }
